# docs/Member api v1.openapi.yaml
openapi: 3.1.0
info:
  title: Gym Member App — Member BFF API
  version: "1.0.0"
  description: >
    Phase-1 contract for the member-facing app. This is the ONLY surface the
    mobile app uses. It lives inside the existing NestJS SaaS as a `/member`
    module, reuses the shared Prisma tenant extension (schema-per-tenant), and
    enforces member-level scoping on top of tenant-schema isolation.

    Auth model: members have no login today. OTP is delivered/verified via the
    existing Supabase IdP (phone OTP). The app exchanges the Supabase token for
    a member access token whose claims carry the tenant schema + memberId.

    Generate from this file:
      - NestJS DTOs / validation (server)
      - Typed API client (React Native app)
servers:
  - url: https://api.yoursaas.com/member/v1
    description: Production
  - url: https://staging-api.yoursaas.com/member/v1
    description: Staging (use fake tenants here for all testing)
security:
  - MemberBearer: []
...
paths:
...
  /nutrition/foods:
    get:
      tags: [Nutrition]
      summary: Search the gym's food catalog
      parameters:
        - name: q
          in: query
          required: false
          schema: { type: string }
          description: Case-insensitive name/brand prefix match. Empty = recent foods.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/FoodSearch" }
  /nutrition/meals:
    post:
      tags: [Nutrition]
      summary: Log a meal with its items (idempotent, offline-safe)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MealLogInput" }
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/MealLogResult" }
  /nutrition/water:
    post:
      tags: [Nutrition]
      summary: Log water intake (idempotent, offline-safe)
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WaterLogInput" }
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/WaterLogResult" }
...
  /community/leaderboard:
    get:
      tags: [Community]
      summary: Gym attendance leaderboard (real check-ins over a period)
      description: >
        Ranks the member's gym by check-ins over the last `period` days. Returns
        first names only (no PII), the current member flagged with `isMe`, plus
        the member's own rank/value. Scoped to the member's tenant.
      parameters:
        - name: period
          in: query
          required: false
          schema: { type: integer, default: 30 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/Leaderboard" }
...
  /community/challenges/{challengeId}/join:
    post:
      tags: [Community]
      summary: Join a challenge (idempotent)
      parameters:
        - name: challengeId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/ChallengeJoinResult" }
        "404":
          $ref: "#/components/responses/NotAMember"
...
  /nutrition/goal:
    put:
      tags: [Nutrition]
      summary: Set the member's daily nutrition goal (upsert)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NutritionGoalInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/NutritionGoal" }

  # ── Health Data Platform (wearable telemetry) ───────────────────────
...
components:
  securitySchemes:
    MemberBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Member access token. Claims MUST include sub=memberId, tenantId
        (the tenant schema), role=member, aud=member. Admin tokens are a
        different audience and are rejected here.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Client-generated key. Server dedupes retries (offline outbox).
  responses:
    NotAMember:
      description: Phone is not linked to any gym membership
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
    Envelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: object }
        meta: { $ref: "#/components/schemas/Meta" }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: MEMBERSHIP_EXPIRED }
            message: { type: string }
            retryable: { type: boolean, default: false }
    Meta:
      type: object
      properties:
        tenantId: { type: string, example: gym_017 }
        serverTime: { type: string, format: date-time }
        cacheTtl: { type: integer, description: Seconds the client may cache }
...
    NutritionGoal:
      type: object
      description: Per-member daily targets.
      properties:
        kcal: { type: integer }
        proteinG: { type: integer }
        carbsG: { type: integer }
        fatG: { type: integer }
        waterMl: { type: integer }
    NutritionGoalInput:
      type: object
      properties:
        kcal: { type: integer }
        proteinG: { type: integer }
        carbsG: { type: integer }
        fatG: { type: integer }
        waterMl: { type: integer }
    FoodItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        brand: { type: string, nullable: true }
        barcode: { type: string, nullable: true }
        servingSize: { type: number }
        servingUnit: { type: string }
        kcal: { type: number }
        proteinG: { type: number }
        carbsG: { type: number }
        fatG: { type: number }
        source: { type: string, enum: [custom, catalog, barcode] }
    FoodSearch:
      type: object
      properties:
        foods:
          type: array
          items: { $ref: "#/components/schemas/FoodItem" }
...
    MealLogItemInput:
      type: object
      required: [name]
      properties:
        foodItemId: { type: string, nullable: true }
        name: { type: string }
        quantity: { type: number }
        unit: { type: string }
        kcal: { type: number }
        proteinG: { type: number }
        carbsG: { type: number }
        fatG: { type: number }
    MealLogInput:
      type: object
      required: [mealType, items]
      properties:
        mealType: { type: string, enum: [breakfast, lunch, dinner, snack] }
        loggedAt: { type: string, format: date-time }
        notes: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/MealLogItemInput" }
    MealLogResult:
      type: object
      properties:
        mealId: { type: string }
    WaterLogInput:
      type: object
      required: [amountMl]
      properties:
        amountMl: { type: integer }
        loggedAt: { type: string, format: date-time }
    WaterLogResult:
      type: object
      properties:
        waterId: { type: string }
        totalMl: { type: integer }

    # ── Health Data Platform (wearable telemetry) ───────────────────
...
    LeaderboardEntry:
      type: object
      properties:
        rank: { type: integer }
        name: { type: string }
        value: { type: integer }
        isMe: { type: boolean }
    Leaderboard:
      type: object
      properties:
        metric: { type: string, example: checkins }
        periodDays: { type: integer }
        entries:
          type: array
          items: { $ref: "#/components/schemas/LeaderboardEntry" }
        myRank: { type: integer, nullable: true }
        myValue: { type: integer }
...
    ChallengeJoinResult:
      type: object
      properties:
        joined: { type: boolean }
        progress: { type: integer }
...